+++ src/env-var/cc-env-var-create.js
@@ -7,6 +7,7 @@ import { css, html, LitElement } from 'lit-element';
 import { dispatchCustomEvent } from '../lib/events.js';
 import { i18n } from '../lib/i18n.js';
 import { defaultThemeStyles } from '../styles/default-theme.js';
+import { linkStyles } from '../templates/cc-link.js';
 
 /**
  * @typedef {import('./types.js').Variable} Variable
@@ -29,6 +30,7 @@ export class CcEnvVarCreate extends LitElement {
   static get properties () {
     return {
       disabled: { type: Boolean },
+      mode: { type: String },
       variablesNames: { type: Array, attribute: 'variables-names' },
       _variableName: { type: String, attribute: false },
       _variableValue: { type: String, attribute: false },
@@ -40,6 +42,9 @@ export class CcEnvVarCreate extends LitElement {
 
     /** @type {boolean} Sets `disabled` attribute on inputs and button. */
     this.disabled = false;
+
+    /** @type {string} Sets the mode of the variables name validation. */
+    this.mode = '';
 
     /** @type {string[]} Sets list of existing variables names (so we can display an error if it already exists). */
     this.variablesNames = [];
@@ -84,9 +89,12 @@ export class CcEnvVarCreate extends LitElement {
 
   render () {
 
-    const isNameInvalid = !validateName(this._variableName);
+    const isNameInvalidSimple = !validateName(this._variableName, 'simple');
+    const isNameInvalidStrict = !validateName(this._variableName, 'strict');
     const isNameAlreadyDefined = this.variablesNames.includes(this._variableName);
-    const hasErrors = isNameInvalid || isNameAlreadyDefined;
+    const hasErrors = (this.mode === 'strict')
+      ? isNameInvalidStrict || isNameAlreadyDefined
+      : isNameInvalidSimple || !isNameInvalidStrict || isNameAlreadyDefined;
 
     return html`
       <cc-flex-gap>
@@ -118,23 +126,32 @@ export class CcEnvVarCreate extends LitElement {
             primary
             ?disabled=${hasErrors || this.disabled}
             @cc-button:click=${this._onSubmit}
-          >${i18n('cc-env-var-create.create-button')}</cc-button>
-
+          >${i18n('cc-env-var-create.create-button')}
+          </cc-button>
         </cc-flex-gap>
       </cc-flex-gap>
 
-      ${(isNameInvalid && this._variableName !== '') ? html`
+      ${(isNameInvalidStrict && this.mode === 'strict' && this._variableName !== '') ? html`
         <cc-error>${i18n(`cc-env-var-create.errors.invalid-name`, { name: this._variableName })}</cc-error>
       ` : ''}
+      ${(isNameInvalidStrict && isNameInvalidSimple && this.mode !== 'strict' && this._variableName !== '') ? html`
+        <cc-error>${i18n(`cc-env-var-create.errors.invalid-name`, { name: this._variableName })}</cc-error>
+      ` : ''}
+      ${(isNameInvalidStrict && !isNameInvalidSimple && this.mode !== 'strict' && this._variableName !== '') ? html`
+        <cc-error notice>${i18n(`cc-env-var-create.info.java-prop`, { name: this._variableName })}</cc-error>
+      ` : ''}
+
       ${isNameAlreadyDefined ? html`
         <cc-error>${i18n(`cc-env-var-create.errors.already-defined-name`, { name: this._variableName })}</cc-error>
       ` : ''}
+
     `;
   }
 
   static get styles () {
     return [
       defaultThemeStyles,
+      linkStyles,
       // language=CSS
       css`
         :host {
@@ -176,6 +193,7 @@ export class CcEnvVarCreate extends LitElement {
       `,
     ];
   }
+
 }
 
 window.customElements.define('cc-env-var-create', CcEnvVarCreate);